requirements.txt
Flask==2.2.5
line-bot-sdk==2.4.2
requests==2.31.0